[fix]:Added Search Functionality and Fixed some bugs
-Added Search Functionality
-Added Guest Login
-Added 404 page
-Fixed the navigation on logout
-Show password fixed on signIn and signUp page

+++ src/components/SignUpForm.jsx
@@ -3,7 +3,6 @@ import "../index.css";
 import { Link,useNavigate } from 'react-router-dom'
 import { publicInstance } from '../utils/axios';
 import { useAuth, useToast } from '../context';
-import Toast from './Toast';
 
 function SignUpForm() {
   const navigate = useNavigate()
@@ -14,6 +13,7 @@ function SignUpForm() {
         confirmPassword:""
     }
   const [signUp,setSignUp] = useState(initialSignUpState)
+  const [showPasswords,setShowPasswords] = useState({password:false,confirmPassword:false})
   const {signInStatusDispatch} = useAuth()
   const {addToast} = useToast()
   const onChangeHandler = (e) => {
@@ -24,27 +24,37 @@ function SignUpForm() {
     return signUp.password === signUp.confirmPassword;
   }
   const onSubmit = async () => {
-    if(checkPasswordMatches()){
-      try{
-        const {status,data} = await publicInstance({
+    if (
+      checkPasswordMatches() &&
+      signUp.fullName !== "" &&
+      signUp.email !== "" &&
+      signUp.password !== "" &&
+      signUp.confirmPassword !== ""
+    ) {
+      try {
+        const { status, data } = await publicInstance({
           method: "post",
           url: "/auth/signup",
           data: {
             ...signUp,
           },
         });
-        if(status===201){
+        if (status === 201) {
           localStorage.setItem("token", data.encodedToken);
           signInStatusDispatch({ type: "SET_USER", payload: data });
-          navigate("/products")
+          navigate("/products");
         }
-      }catch(error){
-        console.log(error)
+      } catch (error) {
+        console.log(error);
+        addToast({
+          type: "Error",
+          msg: "Unable to SignUp",
+        });
       }
-    }else{
+    } else {
       addToast({
         type: "Error",
-        msg: "Password and Confirm Password are not same",
+        msg: "Please fill the SignUp Form Correctly",
       });
     }
   }
@@ -87,10 +97,30 @@ function SignUpForm() {
               onChange={onChangeHandler}
               name="password"
               className="input_regular input_corner"
-              type="password"
+              type={showPasswords.password ? "text" : "password"}
               placeholder="Enter 6 charater or more"
             />
-            <i className="input_icon fas fa-eye"></i>
+            {showPasswords.password ? (
+              <i
+                onClick={() =>
+                  setShowPasswords((prevState) => ({
+                    ...prevState,
+                    password: false,
+                  }))
+                }
+                className="input_icon fas fa-eye-slash"
+              ></i>
+            ) : (
+              <i
+                onClick={() =>
+                  setShowPasswords((prevState) => ({
+                    ...prevState,
+                    password: true,
+                  }))
+                }
+                className="input_icon fas fa-eye"
+              ></i>
+            )}
           </div>
           <div className="input_group">
             <label className="input_label">Confirm Password</label>
@@ -98,10 +128,30 @@ function SignUpForm() {
               onChange={onChangeHandler}
               name="confirmPassword"
               className="input_regular input_corner"
-              type="password"
+              type={showPasswords.confirmPassword ? "text" : "password"}
               placeholder="Enter same as password"
             />
-            <i className="input_icon fas fa-eye-slash"></i>
+            {showPasswords.confirmPassword ? (
+              <i
+                onClick={() =>
+                  setShowPasswords((prevState) => ({
+                    ...prevState,
+                    confirmPassword: false,
+                  }))
+                }
+                className="input_icon fas fa-eye-slash"
+              ></i>
+            ) : (
+              <i
+                onClick={() =>
+                  setShowPasswords((prevState) => ({
+                    ...prevState,
+                    confirmPassword: true,
+                  }))
+                }
+                className="input_icon fas fa-eye"
+              ></i>
+            )}
           </div>
           <div className>
             <button onClick={onSubmit} className="btn btn_primary">
